[#816] Fresh branch for client utils reorg
