+++ client/src/utilities/dates.js
@@ -1,0 +1,14 @@
+/* eslint import/prefer-default-export: "off" */
+
+export function formatDate(date) {
+  let month = date.getMonth() + 1;
+  let day = date.getDate();
+  let hours = date.getHours();
+  let minutes = date.getMinutes();
+  if (month < 10) month = `0${month}`;
+  if (day < 10) day = `0${day}`;
+  if (hours < 10) hours = `0${hours}`;
+  if (minutes < 10) minutes = `0${minutes}`;
+
+  return `${date.getFullYear()}-${month}-${day}`;
+}
