+++ client/src/utilities/validators.js
@@ -1,4 +1,6 @@
-export default function isValidEmail(email) {
+/* eslint import/prefer-default-export: "off" */
+
+export function isValidEmail(email) {
   const regex = /^\w+([.-]?\w+)*@\w+([.-]?\w+)*(\.\w{2,9})+$/;
   return regex.test(email);
 }
